\encoding{utf8} 
\name{hermite-internal}
\alias{hermite-internal}
\alias{int.hermite}
\alias{edg}
\alias{cofi}
\alias{summary.glm.hermite}
\alias{print.summary.glm.hermite}

\title{Internal hermite functions}
\description{Internal hermite functions}
\usage{
int.hermite(x, a, b, m)
edg(y, a, b, m=2)
cofi(p, a, b, m=2)
\method{summary}{glm.hermite}(object, ...)
\method{print.summary}{glm.hermite}(x, ...)
}

\details{These functions are not to be called by the user}
\seealso{
\code{\link{hermite-package}}, \code{\link{dhermite}}, \code{\link{phermite}}, 
\code{\link{qhermite}}, \code{\link{rhermite}}, \code{\link{glm.hermite}}
}
\keyword{internal}
