\name{hi_let}
\docType{data}
\alias{hi_let}
\title{High-LET Radiation Exposure}
\description{
  This data corresponds to an experimental simulation of in vitro
whole body irradiation for high-LET radiation exposure, where
peripheral blood samples were exposed to 10 different doses of
1480MeV oxygen ions. For each dose, the number of dicentrics
chromosomes per blood cell were scored.  
}
\usage{hi_let}
\format{A data frame with 7413 rows and 3 columns.}
\source{DiGiorgio M. et al. (2004) Chromosome aberrations induced in human  
lymphocytes by heavycharged particles in track segment mode. Radiation 
Protection Dosimetry, 108, 47-53.}
\references{
DiGiorgio M., Edwards A. A., Moquet J. E., Finnon P., Hone P. A., Lloyd D. C., 
Kreiner A. J., Schuff J. A., Taja M. R., Vallerga M. B., López F. O. and 
Burlón A., Debray M. E., Valda A. (2004) Chromosome aberrations induced in 
human lymphocytes by heavycharged particles in track segment mode. Radiation 
Protection Dosimetry, 108, 47-53.
}
\keyword{datasets}