\encoding{utf8} 
\name{hermite-internal}
\alias{hermite-internal}
\alias{int.hermite}
\alias{edg}
\alias{cofi}

\title{Internal hermite functions}
\description{Internal hermite functions}
\usage{
int.hermite(x, a, b, m)
edg(y, a, b, m=2)
cofi(p, a, b, m=2)
}

\details{These functions are not to be called by the user}
\seealso{
\code{\link{hermite-package}}, \code{\link{dhermite}}, \code{\link{phermite}}, \code{\link{qhermite}}, \code{\link{rhermite}}
}
\keyword{internal}
